+++ lib/opal/rails/template_handler.rb
@@ -2,21 +2,28 @@ module Opal
   module Rails
     class TemplateHandler
 
-      def self.call(template)
-        new.call(template)
+      def self.call(template, source = template.source)
+        new.call(template, source)
       end
 
-      def call(template)
-        escaped = template.source.gsub(':', '\:')
+      def call(template, source = template.source)
+        escaped = source.gsub(':', '\:')
         string = '%q:' + escaped + ':'
 
         <<-RUBY
+          config = ::Rails.application.config.opal
+
           code = []
           code << 'Object.new.instance_eval {'
-          if ::Rails.application.config.opal.assigns_in_templates
+
+          if config.assign_locals_in_templates?
             code << JSON.parse(local_assigns.to_json).map { |key, val| "\#{key} = \#{val.inspect};" }.join
+          end
+
+          if config.assign_instance_variables_in_templates?
             code << JSON.parse(@_assigns.to_json).map { |key, val| "@\#{key} = \#{val.inspect};" }.join
           end
+
           code << #{string}
           code << '}'
           Opal.compile(code.join("\n"))