+++ spec/integration/assigns_spec.rb
@@ -7,7 +7,7 @@ describe 'controller assignments' do
       Rails.application.config.opal.assigns_in_templates = true
     end
 
-    it 'are in the template' do
+    it 'has them in the template' do
       source = get_source_of '/application/with_assignments.js'
       assignments = opal_eval(source)
 
@@ -29,7 +29,7 @@ describe 'controller assignments' do
       Rails.application.config.opal.assigns_in_templates = false
     end
 
-    it 'are not in the template' do
+    it 'has not them in the template' do
       source = get_source_of '/application/with_assignments.js'
       assignments = opal_eval(source)
       {
@@ -43,6 +43,62 @@ describe 'controller assignments' do
         expect(assignments[ivar]).not_to eq(assignment)
       end
     end
+  end
+
+  context 'when :locals' do
+    before do
+      Rails.application.config.opal.assigns_in_templates = :locals
+    end
+
+    it 'has only locals in the template' do
+      source = get_source_of '/application/with_assignments.js'
+      assignments = opal_eval(source)
+      {
+        :number_var => 1234,
+        :string_var => 'hello',
+        :array_var  => [1,'a'],
+        :hash_var   => {:a => 1, :b => 2},
+        :object_var => {:contents => 'json representation'},
+      }.each_pair do |ivar, assignment|
+        expect(assignments[ivar]).not_to eq(assignment)
+      end
+      {
+        :local_var  => 'i am local',
+      }.each_pair do |ivar, assignment|
+        expect(assignments[ivar]).to eq(assignment)
+      end
+    end
+  end
+
+  context 'when :ivars' do
+    before do
+      Rails.application.config.opal.assigns_in_templates = :ivars
+    end
+
+    it 'has only ivars in the template' do
+      source = get_source_of '/application/with_assignments.js'
+      assignments = opal_eval(source)
+      {
+        :number_var => 1234,
+        :string_var => 'hello',
+        :array_var  => [1,'a'],
+        :hash_var   => {:a => 1, :b => 2},
+        :object_var => {:contents => 'json representation'},
+      }.each_pair do |ivar, assignment|
+        expect(assignments[ivar]).to eq(assignment)
+      end
+      {
+        :local_var  => 'i am local',
+      }.each_pair do |ivar, assignment|
+        expect(assignments[ivar]).not_to eq(assignment)
+      end
+    end
+  end
+
+  it 'has the correct content type' do
+    get '/application/with_assignments.js'
+    expect(response).to be_successful
+    expect(response.headers['Content-Type']).to eq('text/javascript; charset=utf-8')
   end
 
   def get_source_of path